+++ tests/test_ecs.py
@@ -20,304 +20,258 @@
 
 """Unit tests for AWS ECS executor."""
 
+import os
 import tempfile
 from base64 import b64encode
-from unittest.mock import MagicMock
+from pathlib import Path
+from unittest import mock
+from unittest.mock import AsyncMock, MagicMock
 
 import cloudpickle as pickle
 import pytest
 
-from covalent_ecs_plugin.ecs import ECSExecutor
-from covalent_ecs_plugin.scripts import DOCKER_SCRIPT, PYTHON_EXEC_SCRIPT
+from covalent_ecs_plugin.ecs import FUNC_FILENAME, RESULT_FILENAME, ECSExecutor
 
 
-@pytest.fixture
-def ecs_executor(mocker):
-    mocker.patch("covalent_ecs_plugin.ecs.get_config")
-    return ECSExecutor(
-        credentials="mock",
-        profile="mock",
-        s3_bucket_name="mock",
-        ecr_repo_name="mock",
-        ecs_cluster_name="mock",
-        ecs_task_family_name="mock",
-        ecs_task_execution_role_name="mock",
-        ecs_task_role_name="mock",
-        ecs_task_subnet_id="mock",
-        ecs_task_security_group_id="mock",
-        ecs_task_log_group_name="mock",
-        vcpu="mock",
-        memory="mock",
-        poll_freq="mock",
-        cache_dir="mock",
-    )
+class TestECSExecutor:
 
+    MOCK_PROFILE = "my_profile"
+    MOCK_S3_BUCKET_NAME = "s3-bucket"
+    MOCK_ECR_REPO_NAME = "ecs-repo"
+    MOCK_ECS_CLUSTER_NAME = "ecs-cluster"
+    MOCK_ECS_TASK_FAMILY_NAME = "task-family-name"
+    MOCK_ECS_EXECUTION_ROLE = "task-execution-role"
+    MOCK_ECS_TASK_ROLE_NAME = "task-role-name"
+    MOCK_ECS_TASK_SUBNET_ID = "sb-1234"
+    MOCK_ECS_TASK_SG_ID = "sg-123"
+    MOCK_ECS_LOG_GROUP_NAME = "log-group"
+    MOCK_VCPU = 0.1234
+    MOCK_MEMORY = "123"
+    MOCK_POLL_FREQ = 123
+    MOCK_DISPATCH_ID = 112233
+    MOCK_NODE_ID = 1
+    MOCK_TASK_ARN = "task-arn/123"
 
-def test_executor_init_default_values(mocker):
-    """Test that the init values of the executor are set properly."""
-    mocker.patch("covalent_ecs_plugin.ecs.get_config", return_value="mock")
-    mocker.patch("covalent_ecs_plugin.ecs.ECSExecutor._is_valid_subnet_id", return_value=False)
-    mocker.patch(
-        "covalent_ecs_plugin.ecs.ECSExecutor._is_valid_security_group", return_value=False
-    )
-    ecse = ECSExecutor()
-    assert ecse.credentials == "mock"
-    assert ecse.profile == "mock"
-    assert ecse.s3_bucket_name == "mock"
-    assert ecse.ecr_repo_name == "mock"
-    assert ecse.ecs_cluster_name == "mock"
-    assert ecse.ecs_task_family_name == "mock"
-    assert ecse.ecs_task_execution_role_name == "mock"
-    assert ecse.ecs_task_role_name == "mock"
-    assert ecse.ecs_task_subnet_id == "mock"
-    assert ecse.ecs_task_security_group_id == "mock"
-    assert ecse.ecs_task_log_group_name == "mock"
-    assert ecse.vcpu == "mock"
-    assert ecse.memory == "mock"
-    assert ecse.poll_freq == "mock"
-    assert ecse.cache_dir == "mock"
+    @property
+    def MOCK_FUNC_FILENAME(self):
+        return FUNC_FILENAME.format(dispatch_id=self.MOCK_DISPATCH_ID, node_id=self.MOCK_NODE_ID)
 
+    @property
+    def MOCK_RESULT_FILENAME(self):
+        return RESULT_FILENAME.format(dispatch_id=self.MOCK_DISPATCH_ID, node_id=self.MOCK_NODE_ID)
 
-def test_executor_init_validation(mocker):
-    """Test that subnet and security group id is validated."""
-    mocker.patch("covalent_ecs_plugin.ecs.get_config", return_value="mock")
-    mocker.patch("covalent_ecs_plugin.ecs.ECSExecutor._is_valid_subnet_id", return_value=True)
-    mocker.patch(
-        "covalent_ecs_plugin.ecs.ECSExecutor._is_valid_security_group", return_value=False
-    )
+    @property
+    def MOCK_TASK_METADATA(self):
+        return {"dispatch_id": self.MOCK_DISPATCH_ID, "node_id": self.MOCK_NODE_ID}
 
+    @pytest.fixture
+    def mock_executor_config(self, tmp_path):
+        MOCK_CREDENTIALS_FILE: Path = tmp_path / "credentials"
+        MOCK_CREDENTIALS_FILE.touch()
+        config = {
+            "profile": self.MOCK_PROFILE,
+            "s3_bucket_name": self.MOCK_S3_BUCKET_NAME,
+            "ecr_repo_name": self.MOCK_ECR_REPO_NAME,
+            "ecs_cluster_name": self.MOCK_ECS_CLUSTER_NAME,
+            "ecs_task_family_name": self.MOCK_ECS_TASK_FAMILY_NAME,
+            "ecs_task_execution_role_name": self.MOCK_ECS_EXECUTION_ROLE,
+            "ecs_task_role_name": self.MOCK_ECS_TASK_ROLE_NAME,
+            "ecs_task_subnet_id": self.MOCK_ECS_TASK_SUBNET_ID,
+            "ecs_task_security_group_id": self.MOCK_ECS_TASK_SG_ID,
+            "ecs_task_log_group_name": self.MOCK_ECS_LOG_GROUP_NAME,
+            "vcpu": self.MOCK_VCPU,
+            "memory": self.MOCK_MEMORY,
+            "poll_freq": self.MOCK_POLL_FREQ,
+        }
+        return config
 
-def test_is_valid_subnet_id(ecs_executor):
-    """Test the valid subnet checking method."""
-    assert ecs_executor._is_valid_subnet_id("subnet-871545e1") is True
-    assert ecs_executor._is_valid_subnet_id("subnet-871545e") is False
-    assert ecs_executor._is_valid_subnet_id("jlkjlkj871545e1") is False
+    @pytest.fixture
+    def mock_executor(self, mock_executor_config, mocker):
+        # mocker.patch("tempfile")
+        return ECSExecutor(**mock_executor_config)
 
+    @mock.patch.dict(os.environ, {}, clear=True)
+    def test_init_explicit_values(self, mocker, mock_executor_config):
+        """Test executor class values are overridden properly during instantiation"""
 
-def test_is_valid_security_group(ecs_executor):
-    """Test the valid security group checking method."""
-    assert ecs_executor._is_valid_security_group("sg-0043541a") is True
-    assert ecs_executor._is_valid_security_group("sg-0043541") is False
-    assert ecs_executor._is_valid_security_group("80980043541") is False
-
-
-def test_get_aws_account(mocker, ecs_executor):
-    """Test the method to retrieve the aws account."""
-    mm = MagicMock()
-    mocker.patch("covalent_ecs_plugin.ecs.boto3.Session", return_value=mm)
-    ecs_executor._get_aws_account()
-    mm.client().get_caller_identity.called_once_with()
-    mm.client().get_caller_identity.get.called_once_with("Account")
-
-
-def test_execute(mocker, ecs_executor):
-    """Test the execute method."""
-    ecs_executor.vcpu = 1
-    ecs_executor.memory = 1
-
-    def mock_func(x):
-        return x
-
-    mm = MagicMock()
-
-    mocker.patch("covalent_ecs_plugin.ecs.boto3.Session", return_value=mm)
-    package_and_upload_mock = mocker.patch(
-        "covalent_ecs_plugin.ecs.ECSExecutor._package_and_upload"
-    )
-    poll_ecs_task_mock = mocker.patch("covalent_ecs_plugin.ecs.ECSExecutor._poll_ecs_task")
-    query_result_mock = mocker.patch("covalent_ecs_plugin.ecs.ECSExecutor._query_result")
-    ecs_executor.execute(
-        function=mock_func,
-        args=[],
-        kwargs={"x": 1},
-        dispatch_id="mock_dispatch_id",
-        results_dir="/tmp",
-        node_id=1,
-    )
-    package_and_upload_mock.assert_called_once_with(
-        mock_func,
-        "mock_dispatch_id-1",
-        "/tmp/mock_dispatch_id",
-        "result-mock_dispatch_id-1.pkl",
-        [],
-        {"x": 1},
-    )
-    poll_ecs_task_mock.assert_called_once()
-    query_result_mock.assert_called_once()
-    mm.client().register_task_definition.assert_called_once()
-    mm.client().run_task.assert_called_once()
-
-
-def test_format_exec_script(ecs_executor):
-    """Test method that constructs the executable tasks-execution Python script."""
-    kwargs = {
-        "func_filename": "mock_function_filename",
-        "result_filename": "mock_result_filename",
-        "docker_working_dir": "mock_docker_working_dir",
-    }
-    exec_script = ecs_executor._format_exec_script(**kwargs)
-    assert exec_script == PYTHON_EXEC_SCRIPT.format(
-        s3_bucket_name=ecs_executor.s3_bucket_name, **kwargs
-    )
-
-
-def test_format_dockerfile(ecs_executor):
-    """Test method that constructs the dockerfile."""
-    docker_script = ecs_executor._format_dockerfile(
-        exec_script_filename="root/mock_exec_script_filename",
-        docker_working_dir="mock_docker_working_dir",
-    )
-    assert docker_script == DOCKER_SCRIPT.format(
-        func_basename="mock_exec_script_filename", docker_working_dir="mock_docker_working_dir"
-    )
-
-
-def test_upload_file_to_s3(ecs_executor, mocker):
-    """Test method to upload file to s3."""
-    mm = MagicMock()
-    mocker.patch("covalent_ecs_plugin.ecs.boto3.Session", return_value=mm)
-    ecs_executor._upload_file_to_s3(
-        "mock_s3_bucket_name", "mock_temp_function_filename", "mock_s3_function_filename"
-    )
-    mm.client().upload_file.assert_called_once_with(
-        "mock_temp_function_filename", "mock_s3_bucket_name", "mock_s3_function_filename"
-    )
-
-
-def test_ecr_info(ecs_executor, mocker):
-    """Test method to retrieve ecr related info."""
-    mm = MagicMock()
-    mm.client().get_authorization_token.return_value = {
-        "authorizationData": [
-            {
-                "authorizationToken": b64encode(b"fake_token"),
-                "proxyEndpoint": "proxy_endpoint",
-            }
-        ]
-    }
-    mocker.patch("covalent_ecs_plugin.ecs.boto3.Session", return_value=mm)
-    assert ecs_executor._get_ecr_info("mock_image_tag") == (
-        "fake_token",
-        "proxy_endpoint",
-        "proxy_endpoint/mock:mock_image_tag",
-    )
-    mm.client().get_authorization_token.assert_called_once_with()
-
-
-def test_package_and_upload(ecs_executor, mocker):
-    """Test the package and upload method."""
-    upload_file_to_s3_mock = mocker.patch("covalent_ecs_plugin.ecs.ECSExecutor._upload_file_to_s3")
-    format_exec_script_mock = mocker.patch(
-        "covalent_ecs_plugin.ecs.ECSExecutor._format_exec_script", return_value=""
-    )
-    format_dockerfile_mock = mocker.patch(
-        "covalent_ecs_plugin.ecs.ECSExecutor._format_dockerfile", return_value=""
-    )
-    get_ecr_info_mock = mocker.patch(
-        "covalent_ecs_plugin.ecs.ECSExecutor._get_ecr_info",
-        return_value=("", "", ""),
-    )
-    mocker.patch("covalent_ecs_plugin.ecs.shutil.copyfile")
-    mm = MagicMock()
-    tag_mock = MagicMock()
-    mm.images.build.return_value = tag_mock, "logs"
-    mocker.patch("covalent_ecs_plugin.ecs.docker.from_env", return_value=mm)
-
-    ecs_executor._package_and_upload(
-        "mock_transportable_object",
-        "mock_image_tag",
-        "mock_task_results_dir",
-        "mock_result_filename",
-        [],
-        {},
-    )
-    upload_file_to_s3_mock.assert_called_once()
-    format_exec_script_mock.assert_called_once()
-    format_dockerfile_mock.assert_called_once()
-    get_ecr_info_mock.assert_called_once()
-
-
-def test_get_status(ecs_executor):
-    """Test the status checking method."""
-    ecs_mock = MagicMock()
-    ecs_mock.get_paginator().paginate.return_value = []  # Case 1: no tasks found
-    res = ecs_executor.get_status(ecs_mock, "")
-    assert res == ("TASK_NOT_FOUND", -1)
-
-    ecs_mock.get_paginator().paginate.return_value = [
-        {"taskArns": ["mock_task_arn"]}
-    ]  # Case 2 valid task found
-    ecs_mock.describe_tasks.return_value = {
-        "tasks": [
-            {"taskArn": "mock_task_arn", "lastStatus": "RUNNING", "containers": [{"exitCode": 1}]}
-        ]
-    }
-    res = ecs_executor.get_status(ecs_mock, "mock_task_arn")
-    assert res == ("RUNNING", 1)
-
-    ecs_mock.get_paginator().paginate.return_value = [
-        {"taskArns": ["mock_task_arn"]}
-    ]  # Case 3 - task found without any status
-    ecs_mock.describe_tasks.return_value = {
-        "tasks": [{"taskArn": "mock_task_arn", "lastStatus": "FAILED"}]
-    }
-    res = ecs_executor.get_status(ecs_mock, "mock_task_arn")
-    assert res == ("FAILED", -1)
-
-
-def test_poll_ecs_task(mocker, ecs_executor):
-    """Test the method to poll the ecs task."""
-    ecs_executor.poll_freq = 1
-    ecs_mock = MagicMock()
-    time_mock = mocker.patch("covalent_ecs_plugin.ecs.time.sleep")
-    mocker.patch(
-        "covalent_ecs_plugin.ecs.ECSExecutor.get_status",
-        side_effect=[("RUNNING", 1), ("STOPPED", 0), ("RUNNING", 1), ("STOPPED", 2)],
-    )
-    ecs_executor._poll_ecs_task(ecs_mock, "mock_task_arn")
-    time_mock.assert_called_once_with(1)
-
-    with pytest.raises(Exception):
-        ecs_executor._poll_ecs_task(ecs_mock, "mock_task_arn")
-
-
-def test_get_log_events(mocker, ecs_executor):
-    """Test the method to retrieve the log events from the log stream."""
-    logs_client_mock = MagicMock()
-    logs_client_mock.client().get_log_events.return_value = {
-        "events": [{"message": "hello"}, {"message": "world"}]
-    }
-    mocker.patch("covalent_ecs_plugin.ecs.boto3.Session", return_value=logs_client_mock)
-    assert ecs_executor._get_log_events("mock_image_tag", "mock_task_id") == "hello\nworld\n"
-
-
-def test_query_result(mocker, ecs_executor):
-    """Test the method to query the result."""
-    with tempfile.TemporaryDirectory() as temp_dir:
-        with open(f"{temp_dir}/result.pkl", "wb") as f:
-            pickle.dump("mock_result", f)
-        s3_client_mock = MagicMock()
-        mocker.patch("covalent_ecs_plugin.ecs.boto3.Session", return_value=s3_client_mock)
-        logs_mock = mocker.patch(
-            "covalent_ecs_plugin.ecs.ECSExecutor._get_log_events", return_value="mocked_log"
+        # only call to get_config is get_config("executors.ecs.cache_dir")
+        mocker.patch("covalent_ecs_plugin.ecs.get_config", return_value="mock_cache_dir")
+        mocker.patch("covalent_ecs_plugin.ecs.ECSExecutor._is_valid_subnet_id", return_value=True)
+        mocker.patch(
+            "covalent_ecs_plugin.ecs.ECSExecutor._is_valid_security_group", return_value=True
         )
-        assert ecs_executor._query_result(
-            result_filename="result.pkl",
-            task_results_dir=temp_dir,
-            task_arn="mock_task_arn/mock_task_id",
-            image_tag="mock_image_tag",
-        ) == ("mock_result", "mocked_log", "")
-        s3_client_mock.client().download_file.assert_called_once_with(
-            "mock", "result.pkl", f"{temp_dir}/result.pkl"
+
+        executor = ECSExecutor(**mock_executor_config)
+
+        assert executor.profile == self.MOCK_PROFILE
+        assert executor.s3_bucket_name == self.MOCK_S3_BUCKET_NAME
+        assert executor.ecr_repo_name == self.MOCK_ECR_REPO_NAME
+        assert executor.ecs_task_family_name == self.MOCK_ECS_TASK_FAMILY_NAME
+        assert executor.execution_role == self.MOCK_ECS_EXECUTION_ROLE
+        assert executor.ecs_task_role_name == self.MOCK_ECS_TASK_ROLE_NAME
+        assert executor.ecs_task_subnet_id == self.MOCK_ECS_TASK_SUBNET_ID
+        assert executor.ecs_task_security_group_id == self.MOCK_ECS_TASK_SG_ID
+        assert executor.log_group_name == self.MOCK_ECS_LOG_GROUP_NAME
+        assert executor.vcpu == self.MOCK_VCPU
+        assert executor.memory == self.MOCK_MEMORY
+        assert executor.poll_freq == self.MOCK_POLL_FREQ
+
+    @pytest.mark.asyncio
+    async def test_upload_file_to_s3(self, mock_executor, mocker):
+        """Test method to upload file to s3."""
+        boto3_mock = mocker.patch("covalent_ecs_plugin.ecs.boto3")
+
+        def some_function():
+            pass
+
+        await mock_executor._upload_task(
+            some_function, ("some_arg"), {"some": "kwarg"}, self.MOCK_TASK_METADATA
         )
-        logs_mock.assert_called_once_with("mock_image_tag", "mock_task_id")
+        boto3_mock.Session().client().upload_file.assert_called_once()
 
+    def test_is_valid_subnet_id(self, mock_executor):
+        """Test the valid subnet checking method."""
+        assert mock_executor._is_valid_subnet_id("subnet-871545e1") is True
+        assert mock_executor._is_valid_subnet_id("subnet-871545e") is False
+        assert mock_executor._is_valid_subnet_id("jlkjlkj871545e1") is False
 
-def test_cancel(mocker, ecs_executor):
-    """Test the execution cancellation method."""
-    ecs_client_mock = MagicMock()
-    mocker.patch("covalent_ecs_plugin.ecs.boto3.Session", return_value=ecs_client_mock)
-    ecs_executor.cancel("mock_task_arn", "mock_reason")
-    ecs_client_mock.client().stop_task.assert_called_once_with(
-        cluster="mock", task="mock_task_arn", reason="mock_reason"
-    )
+    def test_is_valid_security_group(self, mock_executor):
+        """Test the valid security group checking method."""
+        assert mock_executor._is_valid_security_group("sg-0043541a") is True
+        assert mock_executor._is_valid_security_group("sg-0043541") is False
+        assert mock_executor._is_valid_security_group("80980043541") is False
+
+    @pytest.mark.asyncio
+    async def test_cancel(self, mocker, mock_executor):
+        """Test the execution cancellation method."""
+        boto3_mock = mocker.patch("covalent_ecs_plugin.ecs.boto3")
+        await mock_executor.cancel("mock_task_arn", "mock_reason")
+        boto3_mock.Session().client().stop_task.assert_called_once_with(
+            cluster=self.MOCK_ECS_CLUSTER_NAME, task="mock_task_arn", reason="mock_reason"
+        )
+
+    @pytest.mark.asyncio
+    async def test_query_result(self, mocker, mock_executor, tmp_path: Path):
+        """Test the method to query the result."""
+
+        mock_cwd = tmp_path
+        mock_executor._cwd = mock_cwd.resolve()
+        mock_local_result_path = mock_cwd / self.MOCK_RESULT_FILENAME
+        mock_local_result_path.touch()
+
+        MOCK_RESULT_CONTENTS = "mock_result"
+
+        with open(mock_local_result_path, "wb") as f:
+            pickle.dump(MOCK_RESULT_CONTENTS, f)
+
+        boto3_mock = mocker.patch("covalent_ecs_plugin.ecs.boto3")
+        result = await mock_executor.query_result(task_metadata=self.MOCK_TASK_METADATA)
+        assert result == MOCK_RESULT_CONTENTS
+
+        boto3_mock.Session().client().download_file.assert_called_once_with(
+            self.MOCK_S3_BUCKET_NAME, self.MOCK_RESULT_FILENAME, str(mock_local_result_path)
+        )
+
+    @pytest.mark.asyncio
+    async def test_get_log_events(self, mocker, mock_executor):
+        """Test the method to retrieve the log events from the log stream."""
+
+        boto3_mock = mocker.patch("covalent_ecs_plugin.ecs.boto3")
+        boto3_mock.Session().client().get_log_events.return_value = {
+            "events": [{"message": "hello"}, {"message": "world"}]
+        }
+        log_events = await mock_executor._get_log_events(
+            "task-arn", task_metadata=self.MOCK_TASK_METADATA
+        )
+        assert log_events == "hello\nworld\n"
+
+    @pytest.mark.asyncio
+    async def test_get_status(self, mocker, mock_executor):
+        """Test the status checking method."""
+        boto3_mock = mocker.patch("covalent_ecs_plugin.ecs.boto3")
+        ecs_client_mock = boto3_mock.Session().client()
+
+        # Case 1: no tasks found
+        ecs_client_mock.get_paginator().paginate.return_value = []
+        res = await mock_executor.get_status(self.MOCK_TASK_ARN)
+        assert res == ("TASK_NOT_FOUND", -1)
+
+        # Case 2 valid task found
+        ecs_client_mock.get_paginator().paginate.return_value = [
+            {"taskArns": [self.MOCK_TASK_ARN]}
+        ]
+        ecs_client_mock.describe_tasks.return_value = {
+            "tasks": [
+                {
+                    "taskArn": self.MOCK_TASK_ARN,
+                    "lastStatus": "RUNNING",
+                    "containers": [{"exitCode": 1}],
+                }
+            ]
+        }
+        res = await mock_executor.get_status(self.MOCK_TASK_ARN)
+        assert res == ("RUNNING", 1)
+
+        # Case 3 - task found without any status
+        ecs_client_mock.get_paginator().paginate.return_value = [
+            {"taskArns": [self.MOCK_TASK_ARN]}
+        ]
+        ecs_client_mock.describe_tasks.return_value = {
+            "tasks": [{"taskArn": self.MOCK_TASK_ARN, "lastStatus": "FAILED"}]
+        }
+        res = await mock_executor.get_status(self.MOCK_TASK_ARN)
+        assert res == ("FAILED", -1)
+
+    @pytest.mark.asyncio
+    async def test_poll_ecs_task(self, mocker, mock_executor):
+        """Test the method to poll the ecs task."""
+
+        mock_executor.poll_freq = self.MOCK_POLL_FREQ
+        sleep_mock = mocker.patch("covalent_ecs_plugin.ecs.asyncio.sleep")
+
+        mocker.patch(
+            "covalent_ecs_plugin.ecs.ECSExecutor.get_status",
+            side_effect=[("RUNNING", 1), ("STOPPED", 0)],
+        )
+        await mock_executor._poll_task(self.MOCK_TASK_ARN)
+        sleep_mock.assert_called_once_with(self.MOCK_POLL_FREQ)
+
+        with pytest.raises(Exception):
+            await mock_executor._poll_task(self.MOCK_TASK_ARN)
+
+    @pytest.mark.asyncio
+    async def test_run(self, mocker, mock_executor):
+        """Test the execute method."""
+
+        MOCK_IDENTITY = {"Account": 1234}
+        mock_executor.vcpu = 1
+        mock_executor.memory = 1
+
+        def mock_func(x):
+            return x
+
+        boto3_mock = mocker.patch("covalent_ecs_plugin.ecs.boto3")
+        ecs_client_mock = boto3_mock.Session().client()
+
+        upload_task_mock = mocker.patch("covalent_ecs_plugin.ecs.ECSExecutor._upload_task")
+        validate_credentials_mock = mocker.patch(
+            "covalent_ecs_plugin.ecs.ECSExecutor._validate_credentials"
+        )
+        submit_task_mock = mocker.patch("covalent_ecs_plugin.ecs.ECSExecutor.submit_task")
+        _poll_task_mock = mocker.patch("covalent_ecs_plugin.ecs.ECSExecutor._poll_task")
+        query_result_mock = mocker.patch("covalent_ecs_plugin.ecs.ECSExecutor.query_result")
+
+        validate_credentials_mock.return_value = MOCK_IDENTITY
+
+        await mock_executor.run(
+            function=mock_func, args=[], kwargs={"x": 1}, task_metadata=self.MOCK_TASK_METADATA
+        )
+
+        upload_task_mock.assert_called_once_with(mock_func, [], {"x": 1}, self.MOCK_TASK_METADATA)
+        validate_credentials_mock.assert_called_once()
+        submit_task_mock.assert_called_once_with(self.MOCK_TASK_METADATA, MOCK_IDENTITY)
+
+        returned_task_arn = await submit_task_mock()
+
+        _poll_task_mock.assert_called_once_with(returned_task_arn)
+        query_result_mock.assert_called_once_with(self.MOCK_TASK_METADATA)