+++ CHANGELOG.md
@@ -9,10 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Tests
 
+- Updated tests for ECS Executor now using AWSExecutor base class
+- Added pytest-asyncio
 - Added missing pip deps to functional tests
 
 ### Updated
 
+- ECS Executor now inheriting from AWSExecutor
 - Updated subnet and security group validation to conform to new 17 digit IDs as per <https://aws.amazon.com/about-aws/whats-new/2018/02/longer-format-resource-ids-are-now-available-in-amazon-ec2/>
 
 ## [0.12.0] - 2022-09-01